+++ src/vpic/vpic.cc
@@ -159,7 +159,7 @@ void checkpt_kokkos(vpic_simulation& simulation, const char* fbase)
 #ifndef USE_LEGACY_PARTICLE_ARRAY
     char fname[256];
     FileIO fileIO;
-    int buf_start;
+    size_t buf_start;
     static particle_t * ALIGNED(128) p_buf = NULL;
     if( !p_buf ) MALLOC_ALIGNED( p_buf, PBUF_SIZE, 128 );
     Kokkos::View<particle_t*, Kokkos::HostSpace, Kokkos::MemoryTraits<Kokkos::Unmanaged> > pbuf(p_buf, PBUF_SIZE);
@@ -174,12 +174,12 @@ void checkpt_kokkos(vpic_simulation& simulation, const char* fbase)
         // Copy a PBUF_SIZE hunk of the particle list into the particle buffer,
         // and write it out.  This is simplified from dump_particles since we
         // don't need to call center_p.
-        int bufsize = PBUF_SIZE;
+        size_t bufsize = PBUF_SIZE;
         for( buf_start=0; buf_start<sp->np; buf_start += PBUF_SIZE ) {
             if (buf_start + bufsize > sp->np) bufsize = sp->np - buf_start;
             Kokkos::parallel_for("Populate particle dump buffer",
                     host_execution_policy(0, bufsize),
-                    KOKKOS_LAMBDA (int i) {
+                    KOKKOS_LAMBDA (size_t i) {
 
                     pbuf(i).dx = sp->k_p_h(buf_start + i, particle_var::dx);
                     pbuf(i).dy = sp->k_p_h(buf_start + i, particle_var::dy);
